chore(graph): split metric checkbox component into two for separation of concerns
