+++ src/modules/controls/MetricButton/types.ts
@@ -1,15 +1,15 @@
 import { SleepMetric } from 'modules/controls/MetricConfiguration'
 
-export interface MetricCheckboxProps {
+export interface MetricButtonProps {
   /**
-   * The metric that this checkbox should
-   * toggle for.
+   * The metric that this button should
+   * select for.
    */
   metric: SleepMetric
 
   /**
    * An optional class name to pass
-   * to the underlying checkbox element.
+   * to the underlying button element.
    */
   className?: string
 }
