+++ src/modules/controls/MetricConfiguration/MetricConfiguration.tsx
@@ -7,32 +7,26 @@ export const MetricConfiguration = ({ className }: MetricConfigurationProps) => 
   return (
     <div className={classNames(styles.container, className)}>
       <MetricCheckbox
-        type='checkbox'
         metric={SleepMetric.QUALITY}
       />
 
       <MetricCheckbox
-        type='checkbox'
         metric={SleepMetric.DEEP_SLEEP}
       />
 
       <MetricCheckbox
-        type='checkbox'
         metric={SleepMetric.LIGHT_SLEEP}
       />
 
       <MetricCheckbox
-        type='checkbox'
         metric={SleepMetric.REM_SLEEP}
       />
 
       <MetricCheckbox
-        type='checkbox'
         metric={SleepMetric.AWAKE_TIME}
       />
 
       <MetricCheckbox
-        type='checkbox'
         metric={SleepMetric.DURATION}
       />
     </div>
